refactored tests 'localImage=Default' to be generic
